#25 Agrego dos endpoints para tipo_de_servicio, obtener todos y obtener por id. También agrego los endpoints a la lista de permisos para el Admin

+++ src/main/java/com/cperalta/jardineria/servicios/domain/models/Servicio.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 public class Servicio {
     private Long id;
     private String descripcion;
-    private Long precio;
+    private Float precio;
     private TipoDeServicio tipoDeServicio;
     private Jardinero jardinero;
 }
